fix: convert all tables to lazy, simplify style solution
