+++ libs/datagrid/src/index.ts
@@ -1,4 +1,3 @@
-export * from './lib/ag-grid/ag-grid-dynamic';
 export * from './lib/ag-grid/ag-grid-lazy';
 
 export * from './lib/cells/cumulative-vol-cell';
