+++ libs/trades/src/lib/trades-table.tsx
@@ -5,7 +5,7 @@ import type {
   VegaICellRendererParams,
   VegaValueFormatterParams,
 } from '@vegaprotocol/datagrid';
-import { AgGridDynamic as AgGrid, NumericCell } from '@vegaprotocol/datagrid';
+import { AgGridLazy as AgGrid, NumericCell } from '@vegaprotocol/datagrid';
 import {
   addDecimal,
   addDecimalsFormatNumber,
